Back to 64

+++ src/flag.rs
@@ -92,7 +92,7 @@ pub struct Flag {
     /// LaunchDarkly may affect this flag to prevent poorly performing applications from adversely
     /// affecting upstream service health.
     #[serde(default, skip_serializing_if = "is_default_ratio")]
-    pub sampling_ratio: Option<u32>,
+    pub sampling_ratio: Option<u64>,
 
     /// Determines whether or not this flag will be excluded from the event summarization process.
     ///
@@ -109,7 +109,7 @@ impl Versioned for Flag {
 }
 
 // Used strictly for serialization to determine if a ratio should be included in the JSON.
-fn is_default_ratio(sampling_ratio: &Option<u32>) -> bool {
+fn is_default_ratio(sampling_ratio: &Option<u64>) -> bool {
     sampling_ratio.unwrap_or(1) == 1
 }
 
@@ -129,7 +129,7 @@ pub struct MigrationFlagParameters {
     /// read or write operation. This value can be controlled through the LaunchDarkly UI and
     /// propagated downstream to the SDKs.
     #[serde(skip_serializing_if = "is_default_ratio")]
-    pub check_ratio: Option<u32>,
+    pub check_ratio: Option<u64>,
 }
 
 impl MigrationFlagParameters {
